Fix up complete_reply messages (#8).
The reason that aavogt's complete_reply messages were not showing up is because the "Messagin in IPython" documentation is missing a field for the complete_reply messages. There needs to be a "matched_text" field which contains the part of the line on which completions were computed. I have added that field and it seems to work now.

FYI, I figured this out because tab completion on the notebook was raising javascript errors in the browser. For future reference, this could be a useful thing to check if something in the notebook is not working.