add input code re-broadcast
Our (ipython) docs probably weren't very clear about this, but frontends
expect to get a re-broadcast of the input code, along with an execution
count (since otherwise only the frontend which sent an execute_request
knows what was actually executed). I was working on
ivanov/vim-ipython#86 and found that I wasn't getting the In [ ] prompts
when connected to an IHaskell kernel.

This is my first non-trivial (though still somewhat trivial) Haskell
patch.