Move sandbox setup into initGhci
As it turns out, setSessionDynFlags calls initPackage. Unfortunately,
initPackage only reads the package database the first time it is called.
Therefore, we need to set extraPkgConfs before we call
setSessionDynFlags for the first time in order for the sandbox to be
used. Since initGhci is the first caller of setSessionDynFlags, we make
it responsible for this.