Add automation for easily defining kernels
The module IHaskell.IPython.EasyKernel provides automation for writing
simple IPython kernels. In particular, it provides a record type that
defines configurations and a function that interprets a configuration as
an action in some monad that can do IO.

The configuration consists primarily of functions that implement the
various features of a kernel, such as running code, looking up
documentation, and performing completion. An example for a simple
language that nevertheless has side effects, global state, and timing
effects is included in the examples directory in a file called Calc.hs.

Presently, there is no automation for creating the profile in the
.ipython directory. One should follow the IPython instructions for this.