Fix issues associated with multiple IHaskells installed.
This fix is only a halfway-fix. If there are multiple IHaskell libraries
installed, you can still end up loading an interpreted version of the
wrong one; however, now Display values are serialized to bytestring, so
this will not cause any issues with transferring those between the two
versions. Widget messages are not serializable and cannot be
serializable, so those remain a problem, but instead of crashing it will
instead emit a loud warning on stderr.