Remove utility shell scripts.
All of the shell scripts use 'cabal'; it's long past time to stop
recommending cabal usage and just use stack. (That ship has sailed,
surprisingly quickly!)

Since I am no longer willing to support the cabal workflow it is a
mistake to keep including these scripts; they are likely to be replaced
with newer ones in the future.