Use Aeson.Object for metadata.
The IPython specification for the metadata does not
define that all fields in metadata must be textual.
Latest Jupyters lab uses that feature so the parsing
mechanism became incomplatible with the protocol.
We change Metadata to be an Aeson.Object and introduce
a 'Metadata' newtype in order to keep that explicit
and open for the later changes if needed.

Fixes #912.