labextension npm dependency bounds loosen
Relax upper version bounds on the jupyterlab alpha dependencies
for ihaskell_labextension. It looks to me like these version bounds
were intended to allow the entire major alpha 0.*.* version range, but
the the prefix caret only includes minor versions if the major version
is 0.

Let's set the upper bound to 2.0.0, because according to SemVer, that
is the next version number which we would expect to have breaking
changes.