Add `buffers` field of the Wire Protocol
The Wire Protocol allows for "extra raw data buffer(s)" at the end of a
ZeroMQ message. This commit enables buffers in the ipython-kernel.

https://jupyter-client.readthedocs.io/en/stable/messaging.html#the-wire-protocol

This field has been in "the Wire Protocol" since before the Jupyter Message
specification version 5.0.

https://github.com/jupyter/jupyter_client/issues/520

Resolves issue

https://github.com/gibiansky/IHaskell/issues/1144

I've tested this feature with a proprietary JupyterLab extension and
I've verified that it works. It's difficult to provide a public
reproducible test. The best test may be to get ipywidgets Image working--
that uses buffers.