Switch to MimeCustom with a flag to define how the data is decoded
Rather than having different versions of MimeCustom, have a single
version which encodes the choice of how the conversion to Json
is made. I had planned to make it a function (i.e. have a
`Text -> Value` type) but this would have been more work, given
the instances, so I went with a simple boolean chance. I could
have created my own type, rather than use a boolean, but that
would have meant more work (the instances).

There is still an issue with the Read instance for MimeType,
as we need to define a flag value (at the moment chosen to be
False).