Add a stack-based hie.yaml for haskell-language-server
We want to be able to clone IHaskell, fire up a
haskell-language-server-capable editor, and have it just work with no
extra configuration steps.

If we have Nix and Stack installed on our system, then with this commit
we get what we want. To make this work, we had to add a
special `stack-nix.yaml` which enables Stack Nix integration by default.

This is a bit awkward because now a `stack-nix.yaml.lock` file gets
generated and what should we do with that? Check it in? In this commit
we've decided to ignore it.

See also https://github.com/mpickering/hie-bios/issues/297